pyproject.toml
[project]
name = "mcp-must-gather-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [    
    "fastmcp>=2.8.0",
]
